# contract/openapi.yaml
...
components:
  schemas:
...
    AdminCertificateView:
      properties:
        daysUntilExpiry:
          format: "int32"
          type:
          - "integer"
          - "null"
        domainId:
          format: "uuid"
          type:
          - "string"
          - "null"
        id:
          format: "uuid"
          type: "string"
        kind:
          $ref: "#/components/schemas/CertificateKind"
        lastError:
          type:
          - "string"
          - "null"
        notAfter:
          format: "date-time"
          type:
          - "string"
          - "null"
        scope:
          type: "string"
        status:
          $ref: "#/components/schemas/CertificateStatus"
      required:
      - "id"
      - "kind"
      - "scope"
      - "status"
      type: "object"
...
    CertificateKind:
      enum:
      - "ORIGIN_CA_WILDCARD"
      - "LETS_ENCRYPT"
      type: "string"
    CertificateStatus:
      enum:
      - "ACTIVE"
      - "RENEWING"
      - "FAILED"
      - "REVOKED"
      type: "string"
...
    CreateRequestRequest:
      properties:
        courseOrProject:
          maxLength: 200
          minLength: 0
          type:
          - "string"
          - "null"
        displayName:
          maxLength: 100
          minLength: 0
          type:
          - "string"
          - "null"
        extraNote:
          maxLength: 2000
          minLength: 0
          type:
          - "string"
          - "null"
        orgId:
          format: "uuid"
          type: "string"
        purpose:
          maxLength: 2000
          minLength: 0
          type: "string"
        reqEndDate:
          format: "date"
          type:
          - "string"
          - "null"
        reqStartDate:
          format: "date"
          type:
          - "string"
          - "null"
        type:
          $ref: "#/components/schemas/ResourceType"
        vm:
          anyOf:
          - $ref: "#/components/schemas/CreateVmRequestSpec"
          - type: "null"
        workspaceId:
          format: "uuid"
          type: "string"
      required:
      - "orgId"
      - "purpose"
      - "type"
      - "workspaceId"
      type: "object"
...
    DomainDetailView:
      properties:
        createdAt:
          format: "date-time"
          type: "string"
        fqdn:
          type: "string"
        id:
          format: "uuid"
          type: "string"
        kind:
          $ref: "#/components/schemas/DomainKind"
        releasedAt:
          format: "date-time"
          type:
          - "string"
          - "null"
        reservedUntil:
          format: "date-time"
          type:
          - "string"
          - "null"
        rootDomain:
          type:
          - "string"
          - "null"
        status:
          $ref: "#/components/schemas/DomainStatus"
        verification:
          anyOf:
          - $ref: "#/components/schemas/DomainVerificationView"
          - type: "null"
        verifiedAt:
          format: "date-time"
          type:
          - "string"
          - "null"
        vmId:
          format: "uuid"
          type: "string"
      required:
      - "createdAt"
      - "fqdn"
      - "id"
      - "kind"
      - "status"
      - "vmId"
      type: "object"
    DomainKind:
      enum:
      - "AUTO"
      - "PLATFORM"
      - "CUSTOM"
      type: "string"
    DomainStatus:
      enum:
      - "PENDING"
      - "VERIFYING"
      - "ACTIVE"
      - "FAILED"
      - "REMOVED"
      type: "string"
...
    DomainVerificationView:
      properties:
        aVerified:
          type: "boolean"
        lastCheckedAt:
          format: "date-time"
          type:
          - "string"
          - "null"
        lastError:
          type:
          - "string"
          - "null"
        requiredRecords:
          items:
            $ref: "#/components/schemas/RequiredRecord"
          type: "array"
        token:
          type: "string"
        txtVerified:
          type: "boolean"
      required:
      - "aVerified"
      - "requiredRecords"
      - "token"
      - "txtVerified"
      type: "object"
...
    FieldValidationError:
      description: "검증 실패 필드 한 건 (422 응답의 errors[] 원소)"
      properties:
        field:
          description: "실패한 요청 필드 이름"
          type: "string"
        message:
          description: "한국어 오류 메시지"
          type: "string"
      required:
      - "field"
      - "message"
      type: "object"
...
    MessageResponse:
      properties:
        message:
          type: "string"
      required:
      - "message"
      type: "object"
...
    PageResponseAdminCertificateView:
      properties:
        content:
          items:
            $ref: "#/components/schemas/AdminCertificateView"
          type: "array"
        page:
          format: "int32"
          type: "integer"
        size:
          format: "int32"
          type: "integer"
        totalElements:
          format: "int64"
          type: "integer"
        totalPages:
          format: "int32"
          type: "integer"
      required:
      - "content"
      - "page"
      - "size"
      - "totalElements"
      - "totalPages"
      type: "object"
    PageResponseAdminDomainView:
      properties:
        content:
          items:
            $ref: "#/components/schemas/AdminDomainView"
          type: "array"
        page:
          format: "int32"
          type: "integer"
        size:
          format: "int32"
          type: "integer"
        totalElements:
          format: "int64"
          type: "integer"
        totalPages:
          format: "int32"
          type: "integer"
      required:
      - "content"
      - "page"
      - "size"
      - "totalElements"
      - "totalPages"
      type: "object"
...
    PageResponseAdminRouteView:
      properties:
        content:
          items:
            $ref: "#/components/schemas/AdminRouteView"
          type: "array"
        page:
          format: "int32"
          type: "integer"
        size:
          format: "int32"
          type: "integer"
        totalElements:
          format: "int64"
          type: "integer"
        totalPages:
          format: "int32"
          type: "integer"
      required:
      - "content"
      - "page"
      - "size"
      - "totalElements"
      - "totalPages"
      type: "object"
...
    PageResponseDomainSummaryView:
      properties:
        content:
          items:
            $ref: "#/components/schemas/DomainSummaryView"
          type: "array"
        page:
          format: "int32"
          type: "integer"
        size:
          format: "int32"
          type: "integer"
        totalElements:
          format: "int64"
          type: "integer"
        totalPages:
          format: "int32"
          type: "integer"
      required:
      - "content"
      - "page"
      - "size"
      - "totalElements"
      - "totalPages"
      type: "object"
...
    PageResponseRequestDetailResponse:
      properties:
        content:
          items:
            $ref: "#/components/schemas/RequestDetailResponse"
          type: "array"
        page:
          format: "int32"
          type: "integer"
        size:
          format: "int32"
          type: "integer"
        totalElements:
          format: "int64"
          type: "integer"
        totalPages:
          format: "int32"
          type: "integer"
      required:
      - "content"
      - "page"
      - "size"
      - "totalElements"
      - "totalPages"
      type: "object"
    PageResponseResourceSummaryResponse:
      properties:
        content:
          items:
            $ref: "#/components/schemas/ResourceSummaryResponse"
          type: "array"
        page:
          format: "int32"
          type: "integer"
        size:
          format: "int32"
          type: "integer"
        totalElements:
          format: "int64"
          type: "integer"
        totalPages:
          format: "int32"
          type: "integer"
      required:
      - "content"
      - "page"
      - "size"
      - "totalElements"
      - "totalPages"
      type: "object"
...
    Problem:
      description: "RFC 9457 오류 응답 + Pickle 확장(code, errors). 모든 비정상 응답은 이 형태로 반환됩\
        니다."
      properties:
        code:
          description: "기계 판독용 안정 오류 코드 (클라이언트 분기 기준)"
          type: "string"
        detail:
          description: "이 발생 건에 대한 설명"
          type: "string"
        errors:
          description: "검증 실패 상세 (422에서만 존재)"
          items:
            $ref: "#/components/schemas/FieldValidationError"
          type: "array"
        instance:
          description: "발생 경로"
          type: "string"
        status:
          description: "HTTP 상태 코드"
          format: "int32"
          type: "integer"
        title:
          description: "사람이 읽는 짧은 제목"
          type: "string"
        type:
          description: "문제 유형 URI (기본 about:blank)"
          type: "string"
      required:
      - "title"
      - "status"
      - "code"
      type: "object"
...
    RequestDetailResponse:
      properties:
        courseOrProject:
          type:
          - "string"
          - "null"
        createdAt:
          format: "date-time"
          type: "string"
        displayName:
          type:
          - "string"
          - "null"
        extraNote:
          type:
          - "string"
          - "null"
        id:
          format: "uuid"
          type: "string"
        orgId:
          description: "신청 대상 기관. 행이 사라진 경우에만 null입니다."
          format: "uuid"
          type:
          - "string"
          - "null"
        orgName:
          type: "string"
        purpose:
          type: "string"
        reqEndDate:
          format: "date"
          type:
          - "string"
          - "null"
        reqStartDate:
          format: "date"
          type:
          - "string"
          - "null"
        requesterId:
          description: "신청자. 행이 사라진 경우에만 null입니다."
          format: "uuid"
          type:
          - "string"
          - "null"
        requesterName:
          type: "string"
        review:
          anyOf:
          - $ref: "#/components/schemas/RequestReviewResponse"
          - type: "null"
        status:
          $ref: "#/components/schemas/RequestStatus"
        type:
          $ref: "#/components/schemas/ResourceType"
        updatedAt:
          format: "date-time"
          type: "string"
        vm:
          anyOf:
          - $ref: "#/components/schemas/VmRequestSpecResponse"
          - type: "null"
        workspaceId:
          description: "신청 대상 워크스페이스. 행이 사라진 경우에만 null입니다."
          format: "uuid"
          type:
          - "string"
          - "null"
        workspaceName:
          type: "string"
      required:
      - "createdAt"
      - "id"
      - "orgName"
      - "purpose"
      - "requesterName"
      - "status"
      - "type"
      - "updatedAt"
      - "workspaceName"
      type: "object"
...
    RequestReviewResponse:
      properties:
        comment:
          type:
          - "string"
          - "null"
        decidedAt:
          format: "date-time"
          type: "string"
        decision:
          $ref: "#/components/schemas/ReviewDecision"
        grantedEndDate:
          format: "date"
          type:
          - "string"
          - "null"
        grantedStartDate:
          format: "date"
          type:
          - "string"
          - "null"
        reviewerId:
          description: "결재자. 계정 행이 사라진 경우에만 null입니다."
          format: "uuid"
          type:
          - "string"
          - "null"
        reviewerName:
          type: "string"
      required:
      - "decidedAt"
      - "decision"
      - "reviewerName"
      type: "object"
    RequestStatus:
      enum:
      - "SUBMITTED"
      - "APPROVED"
      - "REJECTED"
      - "CANCELED"
      type: "string"
...
    ResourceType:
      enum:
      - "VM"
      type: "string"
...
    ReviewDecision:
      enum:
      - "APPROVE"
      - "REJECT"
      type: "string"
    RouteStatus:
      enum:
      - "PENDING"
      - "APPLIED"
      - "FAILED"
      - "REMOVED"
      type: "string"
...
  securitySchemes:
    bearerAuth:
      bearerFormat: "JWT"
      description: "액세스 토큰 (JWT HS256, 15분 만료). 클레임: sub(계정 공개 식별자 UUID), role, token_version.\
        \ 비밀번호 변경·계정 비활성화 시 token_version이 올라가 기존 토큰이 즉시 무효화됩니다. 리프레시 토큰은 보안 스킴이 아\
        니라 __Host-pickle_refresh httpOnly 쿠키로만 오갑니다."
      scheme: "bearer"
      type: "http"
info:
  description: "부산대학교 클라우드 플랫폼 Pickle의 REST API. 인증은 JWT Bearer, 오류 응답은 RFC 9457 problem+json(Problem\
    \ 스키마)을 따릅니다."
  title: "Pickle API"
  version: "0.38.0"
openapi: "3.1.0"
paths:
...
  /admin/certificates:
    get:
      operationId: "listAdminCertificates"
      parameters:
      - in: "query"
        name: "orgId"
        required: false
        schema:
          format: "uuid"
          type: "string"
      - in: "query"
        name: "status"
        required: false
        schema:
          $ref: "#/components/schemas/CertificateStatus"
      - in: "query"
        name: "expiringInDays"
        required: false
        schema:
          format: "int32"
          minimum: 1
          type: "integer"
      - in: "query"
        name: "page"
        required: false
        schema:
          default: 0
          format: "int32"
          minimum: 0
          type: "integer"
      - in: "query"
        name: "size"
        required: false
        schema:
          default: 20
          format: "int32"
          maximum: 100
          minimum: 1
          type: "integer"
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseAdminCertificateView"
          description: "OK"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "admin-publishing-controller"
  /admin/domains:
    get:
      operationId: "listAdminDomains"
      parameters:
      - in: "query"
        name: "orgId"
        required: false
        schema:
          format: "uuid"
          type: "string"
      - in: "query"
        name: "kind"
        required: false
        schema:
          $ref: "#/components/schemas/DomainKind"
      - in: "query"
        name: "status"
        required: false
        schema:
          $ref: "#/components/schemas/DomainStatus"
      - in: "query"
        name: "page"
        required: false
        schema:
          default: 0
          format: "int32"
          minimum: 0
          type: "integer"
      - in: "query"
        name: "size"
        required: false
        schema:
          default: 20
          format: "int32"
          maximum: 100
          minimum: 1
          type: "integer"
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseAdminDomainView"
          description: "OK"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "admin-publishing-controller"
  /admin/domains/{domainId}/force-release:
    post:
      operationId: "forceReleaseDomain"
      parameters:
      - in: "path"
        name: "domainId"
        required: true
        schema:
          format: "uuid"
          type: "string"
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: "OK"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "admin-publishing-controller"
  /admin/domains/{domainId}/verify:
    post:
      operationId: "verifyAdminDomain"
      parameters:
      - in: "path"
        name: "domainId"
        required: true
        schema:
          format: "uuid"
          type: "string"
      responses:
        "202":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: "Accepted"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "admin-publishing-controller"
...
  /admin/requests/{requestId}:
    get:
      operationId: "getAdminRequest"
      parameters:
      - in: "path"
        name: "requestId"
        required: true
        schema:
          format: "uuid"
          type: "string"
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RequestDetailResponse"
          description: "OK"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "admin-request-controller"
...
  /admin/routes:
    get:
      operationId: "listAdminRoutes"
      parameters:
      - in: "query"
        name: "orgId"
        required: false
        schema:
          format: "uuid"
          type: "string"
      - in: "query"
        name: "status"
        required: false
        schema:
          $ref: "#/components/schemas/RouteStatus"
      - in: "query"
        name: "page"
        required: false
        schema:
          default: 0
          format: "int32"
          minimum: 0
          type: "integer"
      - in: "query"
        name: "size"
        required: false
        schema:
          default: 20
          format: "int32"
          maximum: 100
          minimum: 1
          type: "integer"
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseAdminRouteView"
          description: "OK"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "admin-publishing-controller"
  /admin/routes/resync:
    post:
      operationId: "resyncRoutes"
      responses:
        "202":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: "Accepted"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "admin-publishing-controller"
  /admin/routes/{routeId}/apply:
    post:
      operationId: "applyAdminRoute"
      parameters:
      - in: "path"
        name: "routeId"
        required: true
        schema:
          format: "uuid"
          type: "string"
      responses:
        "202":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: "Accepted"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "admin-publishing-controller"
...
  /domains:
    get:
      operationId: "listDomains"
      parameters:
      - in: "query"
        name: "vmId"
        required: false
        schema:
          format: "uuid"
          type: "string"
      - in: "query"
        name: "status"
        required: false
        schema:
          $ref: "#/components/schemas/DomainStatus"
      - in: "query"
        name: "page"
        required: false
        schema:
          default: 0
          format: "int32"
          minimum: 0
          type: "integer"
      - in: "query"
        name: "size"
        required: false
        schema:
          default: 20
          format: "int32"
          maximum: 100
          minimum: 1
          type: "integer"
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseDomainSummaryView"
          description: "OK"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "publishing-controller"
...
  /domains/{domainId}/verify:
    post:
      operationId: "verifyDomain"
      parameters:
      - in: "path"
        name: "domainId"
        required: true
        schema:
          format: "uuid"
          type: "string"
      responses:
        "202":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DomainDetailView"
          description: "Accepted"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "publishing-controller"
...
  /requests:
    get:
      operationId: "listRequests"
      parameters:
      - in: "query"
        name: "status"
        required: false
        schema:
          $ref: "#/components/schemas/RequestStatus"
      - in: "query"
        name: "type"
        required: false
        schema:
          $ref: "#/components/schemas/ResourceType"
      - in: "query"
        name: "workspaceId"
        required: false
        schema:
          format: "uuid"
          type: "string"
      - in: "query"
        name: "page"
        required: false
        schema:
          default: 0
          format: "int32"
          minimum: 0
          type: "integer"
      - in: "query"
        name: "size"
        required: false
        schema:
          default: 20
          format: "int32"
          maximum: 100
          minimum: 1
          type: "integer"
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseRequestDetailResponse"
          description: "OK"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "request-controller"
    post:
      operationId: "createRequest"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRequestRequest"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RequestDetailResponse"
          description: "Created"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "request-controller"
  /requests/{requestId}:
    get:
      operationId: "getRequest"
      parameters:
      - in: "path"
        name: "requestId"
        required: true
        schema:
          format: "uuid"
          type: "string"
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RequestDetailResponse"
          description: "OK"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "request-controller"
  /requests/{requestId}/cancel:
    post:
      operationId: "cancelRequest"
      parameters:
      - in: "path"
        name: "requestId"
        required: true
        schema:
          format: "uuid"
          type: "string"
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RequestDetailResponse"
          description: "OK"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "request-controller"
  /resources:
    get:
      operationId: "listResources"
      parameters:
      - in: "query"
        name: "type"
        required: false
        schema:
          $ref: "#/components/schemas/ResourceType"
      - in: "query"
        name: "workspaceId"
        required: false
        schema:
          format: "uuid"
          type: "string"
      - in: "query"
        name: "page"
        required: false
        schema:
          default: 0
          format: "int32"
          minimum: 0
          type: "integer"
      - in: "query"
        name: "size"
        required: false
        schema:
          default: 20
          format: "int32"
          maximum: 100
          minimum: 1
          type: "integer"
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseResourceSummaryResponse"
          description: "OK"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "resource-index-controller"
...
security:
- bearerAuth: []
servers:
- url: "/api/v1"
...